// src/store/user/user.types.ts
// export const USER_ACTION_TYPES = {
//     SET_CURRENT_USER: 'SET_CURRENT_USER',
//     CHECK_USER_AUTHENTIFICATION: 'CHECK_USER_AUTHENTIFICATION',
//     USER_SIGIN_SUCCESS: 'USER_SIGIN_SUCCESS',
//     USER_SIGNIN_FAILed: 'USER_SIGNIN_FAIL',
//     USER_SIGIN_GOOGLE_POPUP: 'USER_SIGIN_GOOGLE_POPUP',
//     USER_SIGNOUT: 'USER_SIGNOUT',
//     USER_SIGNOUT_SUCCESS: 'USER_SIGNOUT_SUCCESS',
//     USER_SIGNOUT_FAILED: 'USER_SIGNOUT_FAILED',
//     USER_SIGIN_EMAIL: 'USER_SIGIN_EMAIL',
//     USER_SIGNUP: 'USER_SIGNUP',
//     USER_SIGNUP_FAILED: 'USER_SIGNUP_FAILED',
// }


export enum USER_ACTION_TYPES {
    SET_CURRENT_USER = 'SET_CURRENT_USER',
    CHECK_USER_AUTHENTIFICATION = 'CHECK_USER_AUTHENTIFICATION',
    USER_SIGIN_SUCCESS =  'USER_SIGIN_SUCCESS',
    USER_SIGNIN_FAILed =  'USER_SIGNIN_FAIL',
    USER_SIGIN_GOOGLE_POPUP=  'USER_SIGIN_GOOGLE_POPUP',
    USER_SIGNOUT = 'USER_SIGNOUT',
    USER_SIGNOUT_SUCCESS = 'USER_SIGNOUT_SUCCESS',
    USER_SIGNOUT_FAILED =  'USER_SIGNOUT_FAILED',
    USER_SIGIN_EMAIL = 'USER_SIGIN_EMAIL',
    USER_SIGNUP =  'USER_SIGNUP',
    USER_SIGNUP_FAILED = 'USER_SIGNUP_FAILED',
}